+++ src/main/resources/db/migrations/V1.3.1__create_loan_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE loan
     id bigint AUTO_INCREMENT PRIMARY KEY,
     amount_granted integer NOT NULL,
     amount_to_repay float NOT NULL,
-    amount_payed integer NOT NULL,
+    amount_payed float NOT NULL,
     customer_id bigint NOT NULL,
 
     FOREIGN KEY (customer_id) REFERENCES customer (id)
